Moved the stripe files to /billing
